requirements.txt
# Azure Functions dependencies
azure-functions>=1.15.0
azure-storage-blob>=12.19.0  # Updated from 12.9.0
opencensus-ext-azure>=1.1.9
azure-identity>=1.13.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.2  # Updated from 3.0.9

# Environment variables
python-dotenv>=1.0.0  # Updated from 0.19.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.3.0
flake8>=6.0.0
openai>=1.3.0

# Web server
flask>=3.0.0
gunicorn>=21.2.0  # Updated from 20.1.0
werkzeug>=3.0.1